PRO-5526: Move decode function to service
